# lib/api-zod/src/generated/types/feeAdminState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeeAuditEntry } from './feeAuditEntry';
import type { FeeSettings } from './feeSettings';

export interface FeeAdminState {
  settings: FeeSettings;
  audit: FeeAuditEntry[];
  appliesTo: string;
  processingChargeNotice: string;
}